backend/requirements.txt
flask==3.0.3
flask-cors==4.0.0
requests==2.31.0
python-dotenv==1.0.1
supabase==2.27.2
pyjwt==2.10.1
schedule==1.2.0